requirements.txt
Authlib==1.2.0
autopep8==2.0.2
cachetools==5.3.0
cairocffi==1.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
decorator==4.4.2
dropbox==11.36.0
fonttools==4.38.0
gizeh==0.1.11
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.2
google-cloud-logging==3.5.0
google-cloud-speech==2.17.3
google-cloud-storage==2.7.0
google-cloud-texttospeech==2.14.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.51.1
gTTS==2.3.1
httplib2==0.21.0
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
importlib-resources==6.0.1
kiwisolver==1.4.4
matplotlib==3.7.0
moviepy==1.0.3
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
ply==3.11
proglog==0.1.10
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==12.1.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
srt==3.5.2
stone==3.3.1
tomli==2.0.1
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.14
Wave==0.0.2
zipp==3.16.2
